added some changes added finding direct path if AllowedTransfer is False

+++ src/bg/sofia/uni/fmi/mjt/itinerary/City.java
@@ -1,11 +1,13 @@
 package bg.sofia.uni.fmi.mjt.itinerary;
 
+import java.math.BigDecimal;
+
 public record City(String name, Location location) {
 
     public static final int AVERAGE_PRICE_BY_KM = 20;
 
-    public long calculateHeuristic(City to) {
-        return location.calculateManhattanDistance(to.location) * AVERAGE_PRICE_BY_KM;
+    public BigDecimal calculateHeuristic(City to) {
+        return BigDecimal.valueOf(location.calculateManhattanDistance(to.location) * AVERAGE_PRICE_BY_KM);
     }
 
 }
